dotfiles: used --git-dir and --work-tree to move from $HOME
Moved .git folder to .dotfiles/.git to avoid making home directory a git
repo. Reason being this was tracked by my neovim setup, which was
looking for git root folder to tell whether something is a project or
not. Since `everything` in $HOME was a project up until then, nvim-tree
was providing whole home as tree of the project, which was not desired.

As Archlinux wiki suggests here, i used --git-dir and added dot alias:
https://wiki.archlinux.org/title/Dotfiles